# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: New Model Bread - Bakery Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Successful login
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # Users
...
  /branches:
    get:
      operationId: listBranches
      tags: [branches]
      summary: List branches
      responses:
        "200":
          description: List of branches
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Branch"
    post:
      operationId: createBranch
      tags: [branches]
      summary: Create a branch
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBranchBody"
      responses:
        "201":
          description: Created branch
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Branch"
  /branches/{id}:
    patch:
      operationId: updateBranch
      tags: [branches]
      summary: Update branch
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateBranchBody"
      responses:
        "200":
          description: Updated branch
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Branch"

  # Production
...
  /inventory:
    get:
      operationId: listInventory
      tags: [inventory]
      summary: List inventory items
      parameters:
        - name: branchId
          in: query
          required: false
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: List of inventory items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/InventoryItem"
    post:
      operationId: createInventoryItem
      tags: [inventory]
      summary: Create inventory item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInventoryItemBody"
      responses:
        "201":
          description: Created inventory item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryItem"
  /inventory/{id}:
    get:
      operationId: getInventoryItem
      tags: [inventory]
      summary: Get inventory item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Inventory item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryItem"
    patch:
      operationId: updateInventoryItem
      tags: [inventory]
      summary: Update inventory item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateInventoryItemBody"
      responses:
        "200":
          description: Updated inventory item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryItem"
    delete:
      operationId: deleteInventoryItem
      tags: [inventory]
      summary: Delete inventory item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /inventory/{id}/adjust:
    post:
      operationId: adjustInventory
      tags: [inventory]
      summary: Adjust inventory quantity
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdjustInventoryBody"
      responses:
        "200":
          description: Adjusted inventory item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryItem"
  /inventory/low-stock:
    get:
      operationId: getLowStockItems
      tags: [inventory]
      summary: Get low stock items
      parameters:
        - name: branchId
          in: query
          required: false
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: Low stock inventory items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/InventoryItem"

  # Sales
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success

    # Auth schemas
    LoginBody:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      required:
        - username
        - password
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
      required:
        - token
        - user

    # User schemas
    User:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        fullName:
          type: string
        email:
          type: ["string", "null"]
        role:
          type: string
          enum: [managing_director, manager, receptionist, production_staff]
        branchId:
          type: ["integer", "null"]
        isActive:
          type: boolean
        createdAt:
          type: string
        updatedAt:
          type: string
        deletedAt:
          type: ["string", "null"]
      required:
        - id
        - username
        - fullName
        - role
        - isActive
        - createdAt
        - updatedAt
...
    Branch:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        location:
          type: ["string", "null"]
        isActive:
          type: boolean
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - name
        - isActive
        - createdAt
        - updatedAt
    CreateBranchBody:
      type: object
      properties:
        name:
          type: string
        location:
          type: ["string", "null"]
      required:
        - name
    UpdateBranchBody:
      type: object
      properties:
        name:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        isActive:
          type: ["boolean", "null"]

    # Production schemas
...
    InventoryItem:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        category:
          type: string
        unit:
          type: string
        currentQuantity:
          type: number
        minimumQuantity:
          type: number
        costPerUnit:
          type: number
        branchId:
          type: integer
        branchName:
          type: string
        isLowStock:
          type: boolean
        createdAt:
          type: string
        updatedAt:
          type: string
        deletedAt:
          type: ["string", "null"]
      required:
        - id
        - name
        - category
        - unit
        - currentQuantity
        - minimumQuantity
        - costPerUnit
        - branchId
        - branchName
        - isLowStock
        - createdAt
        - updatedAt
    CreateInventoryItemBody:
      type: object
      properties:
        name:
          type: string
        category:
          type: string
        unit:
          type: string
        currentQuantity:
          type: number
        minimumQuantity:
          type: number
        costPerUnit:
          type: number
        branchId:
          type: integer
      required:
        - name
        - category
        - unit
        - currentQuantity
        - minimumQuantity
        - costPerUnit
        - branchId
    UpdateInventoryItemBody:
      type: object
      properties:
        name:
          type: ["string", "null"]
        category:
          type: ["string", "null"]
        unit:
          type: ["string", "null"]
        currentQuantity:
          type: ["number", "null"]
        minimumQuantity:
          type: ["number", "null"]
        costPerUnit:
          type: ["number", "null"]
    AdjustInventoryBody:
      type: object
      properties:
        adjustment:
          type: number
        reason:
          type: string
      required:
        - adjustment
        - reason

    # Sales schemas
...