# lib/api-zod/src/generated/types/createInventoryItemBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * New Model Bread - Bakery Management System API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateInventoryItemBody {
  name: string;
  category: string;
  unit: string;
  currentQuantity: number;
  minimumQuantity: number;
  costPerUnit: number;
  branchId: number;
}