# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * New Model Bread - Bakery Management System API
 * OpenAPI spec version: 0.1.0
 */
import type { UserRole } from "./userRole";

export interface User {
  id: number;
  username: string;
  fullName: string;
  /** @nullable */
  email?: string | null;
  role: UserRole;
  /** @nullable */
  branchId?: number | null;
  isActive: boolean;
  createdAt: string;
  updatedAt: string;
  /** @nullable */
  deletedAt?: string | null;
}